requirements.txt
opentelemetry.api
opentelemetry.sdk
opentelemetry.exporter.otlp.proto.grpc
opentelemetry.exporter.otlp.proto.http
opentelemetry.instrumentation.logging
pyrfc3339
ghapi
requests
python-dateutil


